@@ -3,7 +3,7 @@ description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -18,50 +18,43 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
+  cupertino_icons: ^1.0.2
+
   flutter:
     sdk: flutter
-  intl: ^0.17.0
-  provider: ^5.0.0
-  sqflite: ^2.0.0+3
+
+  flutter_local_notifications: ^13.0.0
+  flutter_native_timezone: ^1.0.10
+  intl: ^0.18.0
   path: ^1.8.0
   path_provider: ^2.0.2
-  table_calendar: ^3.0.0
-  syncfusion_flutter_charts: ^19.1.69+1
-  shared_preferences: ^2.0.6
   permission_handler: ^8.3.0
-  file_picker: ^4.3.0
-  dart_code_metrics: ^4.21.3
-
-  # animated_splash_screen: ^1.1.0
-  flutter_local_notifications: ^6.0.0
-  url_launcher: ^6.0.8
+  provider: ^6.1.1
   rxdart: ^0.26.0-nullsafety.1
-  flutter_native_timezone: ^1.0.10
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  shared_preferences: ^2.0.6
+  sqflite: ^2.0.0+3
+  syncfusion_flutter_charts: ^23.1.44
+  table_calendar: ^3.0.9
+  url_launcher: ^6.1.1
+  file_picker: ^5.5.0
+  fl_chart: ^0.61.0
+  font_awesome_flutter: ^10.6.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-
-
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
-
 # The following section is specific to Flutter.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
   uses-material-design: true
-
 
   # To add assets to your application, add an assets section, like this:
   assets:
@@ -73,15 +66,11 @@ flutter:
     - assets/icon/github.png
     - assets/icon/linkedin.png
     - assets/icon/graph.png
-    
     # - images/a_dot_ham.jpeg
-
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
-
   # For details regarding adding assets from package dependencies, see
   # https://flutter.dev/assets-and-images/#from-packages
-
   # To add custom fonts to your application, add a fonts section here,
   # in this "flutter" section. Each entry in this list should have a
   # "family" key with the font family name, and a "fonts" key with a
